refactor: 把`types.rs`拆分成多个文件
