+++ src-tauri/src/types/mod.rs
@@ -1,0 +1,7 @@
+mod comic;
+mod qrcode_data;
+mod qrcode_status;
+
+pub use comic::*;
+pub use qrcode_data::*;
+pub use qrcode_status::*;
