+++ src-tauri/src/types/qrcode_data.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+use specta::Type;
+
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize, Type)]
+#[serde(rename_all = "camelCase")]
+pub struct QrcodeData {
+    pub base64: String,
+    #[serde(rename = "auth_code")]
+    pub auth_code: String,
+}
